requirements.txt
qrcode[pil]==8.2
reportlab==4.4.4
Pillow==11.3.0
opencv-python==4.12.0.88
pyzbar==0.1.9
pdf2image==1.17.0
